proceso ej post/put

+++ src/main/java/com/carlos/hogwarts/model/Casa.java
@@ -1,11 +1,20 @@
 package com.carlos.hogwarts.model;
 
-import jakarta.persistence.*;
-import lombok.NoArgsConstructor;
-import lombok.Data;
 import java.util.List;
 
 import com.fasterxml.jackson.annotation.JsonManagedReference;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
 @NoArgsConstructor 
@@ -15,7 +24,7 @@ import com.fasterxml.jackson.annotation.JsonManagedReference;
 public class Casa {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id_casa;
+    private Long id_casa;
 
     @Column(name = "nombre")
     private String nombre;
